VilkårsvurderingDto -> VilkårDto VilkårDto -> VilkårsvurderingDto
DelvilkårsvurderingDto -> DelvilkårDto
