+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/dto/VilkårsvurderingDto.kt
@@ -1,77 +1,14 @@
 package no.nav.tilleggsstonader.sak.vilkår.dto
 
-import no.nav.tilleggsstonader.sak.vilkår.domain.Delvilkårsvurdering
-import no.nav.tilleggsstonader.sak.vilkår.domain.VilkårType
-import no.nav.tilleggsstonader.sak.vilkår.domain.Vilkårsresultat
-import no.nav.tilleggsstonader.sak.vilkår.domain.Vilkårsvurdering
-import no.nav.tilleggsstonader.sak.vilkår.domain.Vurdering
-import no.nav.tilleggsstonader.sak.vilkår.regler.RegelId
-import no.nav.tilleggsstonader.sak.vilkår.regler.SvarId
-import java.time.LocalDateTime
-import java.util.UUID
-
 data class VilkårsvurderingDto(
-    val id: UUID,
-    val behandlingId: UUID,
-    val resultat: Vilkårsresultat,
-    val vilkårType: VilkårType,
-    val barnId: UUID? = null,
-    val endretAv: String,
-    val endretTid: LocalDateTime,
-    val delvilkårsett: List<DelvilkårsvurderingDto> = emptyList(),
-    val opphavsvilkår: OpphavsvilkårDto?,
+    val vilkårsett: List<VilkårDto>,
+    val grunnlag: VilkårGrunnlagDto,
 )
 
-data class OpphavsvilkårDto(
-    val behandlingId: UUID,
-    val endretTid: LocalDateTime,
+data class VilkårGrunnlagDto(
+    val fellesgrunnlag: Fellesgrunnlag,
 )
 
-data class OppdaterVilkårsvurderingDto(val id: UUID, val behandlingId: UUID)
-
-data class SvarPåVurderingerDto(
-    val id: UUID,
-    val behandlingId: UUID,
-    val delvilkårsvurderinger: List<DelvilkårsvurderingDto>,
+data class Fellesgrunnlag(
+    val navn: String,
 )
-
-data class GjenbrukVilkårsvurderingerDto(val behandlingId: UUID, val kopierBehandlingId: UUID)
-
-data class DelvilkårsvurderingDto(
-    val resultat: Vilkårsresultat,
-    val vurderinger: List<VurderingDto>,
-) {
-
-    /**
-     * @return regelId for første svaret som er hovedregeln på delvilkåret
-     */
-    fun hovedregel() = this.vurderinger.first().regelId
-}
-
-data class VurderingDto(
-    val regelId: RegelId,
-    val svar: SvarId? = null,
-    val begrunnelse: String? = null,
-)
-
-fun Vurdering.tilDto() = VurderingDto(this.regelId, this.svar, this.begrunnelse)
-
-fun Delvilkårsvurdering.tilDto() = DelvilkårsvurderingDto(this.resultat, this.vurderinger.map { it.tilDto() })
-
-fun Vilkårsvurdering.tilDto() =
-    VilkårsvurderingDto(
-        id = this.id,
-        behandlingId = this.behandlingId,
-        resultat = this.resultat,
-        vilkårType = this.type,
-        barnId = this.barnId,
-        endretAv = this.sporbar.endret.endretAv,
-        endretTid = this.sporbar.endret.endretTid,
-        delvilkårsett = this.delvilkårsvurdering.delvilkårsvurderinger
-            .filter { it.resultat != Vilkårsresultat.IKKE_AKTUELL }
-            .map { it.tilDto() },
-        opphavsvilkår = this.opphavsvilkår?.let { OpphavsvilkårDto(it.behandlingId, it.vurderingstidspunkt) },
-    )
-
-fun DelvilkårsvurderingDto.svarTilDomene() = this.vurderinger.map { it.tilDomene() }
-fun VurderingDto.tilDomene() = Vurdering(this.regelId, this.svar, this.begrunnelse)
